Ultimos ajustes y documentacion completa

+++ src/main/java/ar/com/proyectosemillas/Quinoa.java
@@ -2,17 +2,29 @@ package ar.com.proyectosemillas;
 
 import java.util.List;
 import java.util.ListIterator;
-
+/**
+ * Clase hija de Planta.
+ * 
+ * @author Sol Noguera
+ * @see Planta
+ *
+ */
 public class Quinoa extends Planta {
 	
 	private int horasBajoSol;
 	
-	public Quinoa(int anioObtencion, float altura, int horasBajoSol) {
+	/**
+	 * Constructor unico de la clase Quinoa
+	 * @param anioObtencion
+	 * @param altura
+	 * @param horasBajoSol
+	 */
+	public Quinoa(int anioObtencion, double altura, int horasBajoSol) {
 		super(anioObtencion, altura);
 		this.horasBajoSol = horasBajoSol;
 	}
 	
-	public float getHorasBajoSol() {
+	public int getHorasBajoSol() {
 		return this.horasBajoSol;
 	}
